+++ mycodo/inputs/atlas_flow.py
@@ -201,7 +201,7 @@ INPUT_INFORMATION = {
         {
             'id': 'clear_total_volume',
             'type': 'button',
-            'name': lazy_gettext('Clear Total Volume')
+            'name': f"{lazy_gettext('Clear Total')}: {lazy_gettext('Volume')}"
         },
         {
             'type': 'message',